Crate som_interpreter_bc

This is the interpreter for the Simple Object Machine.

Modules

compiler

Facilities for compiling code into bytecode. Compiles parser AST to bytecode. This module only needs to expose the compile_class() function: the rest of the VM should not need access to more than that, barring testing.

debug

Used for debugging. Debugging facilities.

gc

Structs and info related to interacting with the GC

hashcode

Facilities for manipulating values.

interpreter

The interpreter’s main data structure.

primitives

Definitions for all supported primitives.

universe

The collection of all known SOM objects during execution.

value

Facilities for manipulating values.

vm_objects

VM objects.

Macros

pop_args_from_stack

stack_drop

stack_last

stack_n_last_elements

stack_nth_back

stack_pop

stack_push

stack_truncate

Statics

INTERPRETER_RAW_PTR_CONST

See UNIVERSE_RAW_PTR_CONST.

UNIVERSE_RAW_PTR_CONST

Raw pointer needed to trace GC roots. Meant to be accessed only non-mutably, hence the “CONST” in the name.