som_interpreter_bc::interpreter

Struct Interpreter

pub struct Interpreter {
    pub start_time: Instant,
    pub bytecode_idx: u16,
    pub current_frame: UnsafeCell<Gc<Frame>>,
    pub current_bytecodes: *const Vec<Bytecode>,
    pub frame_method_root: Gc<Method>,
    pub frame_args_root: Option<Vec<Value>>,
}

Fields
start_time: Instant

The time record of the interpreter’s creation.

bytecode_idx: u16

The current bytecode index.

current_frame: UnsafeCell<Gc<Frame>>

The current frame.

current_bytecodes: *const Vec<Bytecode>

Pointer to the frame’s bytecodes, to not have to read them from the frame directly

frame_method_root: Gc<Method>

GC can trigger when the interpreter wants to allocate a new frame. We’re then in a situation where we’ve looked up a Method (which is how we knew we were dealing with a non-primitive, and so that we had to create a frame) So this method can’t be stored on the Rust stack, or GC would miss it. Therefore: we keep it reachable there.

frame_args_root: Option<Vec<Value>>
Implementations

impl Interpreter

pub fn new(base_frame: Gc<Frame>) -> Self

pub fn get_current_frame(&self) -> Gc<Frame>

Return the current frame. It’s in an UnsafeCell for moving GC reasons: you get many bugs by using Gc by itself, since Rust assumes that it hasn’t moved when it in fact very much has

pub fn get_current_frame_mut(&mut self) -> &mut Gc<Frame>

pub fn push_method_frame( &mut self, method: Gc<Method>, nbr_args: usize, mutator: &mut GCInterface, ) -> Gc<Frame>

Creates and allocates a new frame corresponding to a method. nbr_args is the number of arguments, including the self value, which it takes from the previous frame.

pub fn push_method_frame_with_args( &mut self, method: Gc<Method>, args: Vec<Value>, mutator: &mut GCInterface, ) -> Gc<Frame>

Creates and allocates a new frame corresponding to a method, with arguments provided. Used in primitives and corner cases like DNU calls.

pub fn push_block_frame( &mut self, nbr_args: usize, mutator: &mut GCInterface, ) -> Gc<Frame>

Creates and allocates a new frame corresponding to a method.

pub fn pop_frame(&mut self)

pub fn pop_n_frames(&mut self, n: u8)

pub fn run(&mut self, universe: &mut Universe) -> Option<Value>