Struct som_interpreter_bc::universe::Universe

pub struct Universe {
    pub interner: Interner,
    pub globals: HashMap<Interned, Value>,
    pub classpath: Vec<PathBuf>,
    pub core: CoreClasses,
}

The central data structure for the interpreter.

It represents the complete state of the interpreter, like the known class definitions, the string interner and the stack frames.

Fields
interner: Interner

The string interner for symbols.

globals: HashMap<Interned, Value>

The known global bindings.

classpath: Vec<PathBuf>

The path to search in for new classes.

core: CoreClasses

The interpreter's core classes.

Implementations

impl Universe

pub fn with_classpath(classpath: Vec<PathBuf>) -> Result<Self, Error>

Initialize the universe from the given classpath.

pub fn load_system_class(
interner: &mut Interner,
classpath: &[impl AsRef<Path>],
class_name: impl Into<String>
) -> Result<SOMRef<Class>, Error>

Load a system class (with an incomplete hierarchy).

pub fn load_class(
&mut self,
class_name: impl Into<String>
) -> Result<SOMRef<Class>, Error>

Load a class from its name into this universe.

pub fn load_class_from_path(
&mut self,
path: impl AsRef<Path>
) -> Result<SOMRef<Class>, Error>

Load a class from its path into this universe.

pub fn nil_class(&self) -> SOMRef<Class>

Get the Nil class.

pub fn system_class(&self) -> SOMRef<Class>

Get the System class.

pub fn object_class(&self) -> SOMRef<Class>

Get the Object class.

pub fn symbol_class(&self) -> SOMRef<Class>

Get the Symbol class.

pub fn string_class(&self) -> SOMRef<Class>

Get the String class.

pub fn array_class(&self) -> SOMRef<Class>

Get the Array class.

pub fn integer_class(&self) -> SOMRef<Class>

Get the Integer class.

pub fn double_class(&self) -> SOMRef<Class>

Get the Double class.

pub fn block_class(&self) -> SOMRef<Class>

Get the Block class.

pub fn block1_class(&self) -> SOMRef<Class>

Get the Block1 class.

pub fn block2_class(&self) -> SOMRef<Class>

Get the Block2 class.

pub fn block3_class(&self) -> SOMRef<Class>

Get the Block3 class.

pub fn true_class(&self) -> SOMRef<Class>

Get the True class.

pub fn false_class(&self) -> SOMRef<Class>

Get the False class.

pub fn metaclass_class(&self) -> SOMRef<Class>

Get the Metaclass class.

pub fn method_class(&self) -> SOMRef<Class>

Get the Method class.

pub fn primitive_class(&self) -> SOMRef<Class>

Get the Primitive class.

impl Universe

pub fn intern_symbol(&mut self, symbol: &str) -> Interned

Intern a symbol.

pub fn lookup_symbol(&self, symbol: Interned) -> &str

Lookup a symbol.

pub fn lookup_global(&self, idx: Interned) -> Option<Value>

Search for a global binding.

pub fn assign_global(&mut self, name: Interned, value: Value) -> Option<()>

Assign a value to a global binding.

impl Universe

pub fn escaped_block(
&mut self,
interpreter: &mut Interpreter,
value: Value,
block: Rc<Block>
) -> Option<()>

Call escapedBlock: on the given value, if it is defined.

pub fn does_not_understand(
&mut self,
interpreter: &mut Interpreter,
value: Value,
symbol: Interned,
args: Vec<Value>
) -> Option<()>

Call doesNotUnderstand: on the given value, if it is defined.

pub fn unknown_global(
&mut self,
interpreter: &mut Interpreter,
value: Value,
name: Interned
) -> Option<()>

Call unknownGlobal: on the given value, if it is defined.

pub fn initialize(
&mut self,
interpreter: &mut Interpreter,
args: Vec<Value>
) -> Option<()>

Call System>>#initialize: with the given name, if it is defined.