som_interpreter_bc::value

Struct Value

#[repr(transparent)]
pub struct Value(pub BaseValue);

Represents an SOM value.

Tuple Fields
0: BaseValue
Implementations

impl Value

pub const TRUE: Self

pub const FALSE: Self

pub const NIL: Self

pub const INTEGER_ZERO: Self

pub const INTEGER_ONE: Self

pub fn new_boolean(value: bool) -> Self

pub fn new_integer(value: i32) -> Self

pub fn new_double(value: f64) -> Self

pub fn new_symbol(value: Interned) -> Self

pub fn new_char(value: char) -> Self

pub fn new_big_integer(value: Gc<BigInt>) -> Self

pub fn new_string(value: Gc<String>) -> Self

pub fn Boolean(value: bool) -> Self

pub fn Char(value: char) -> Self

pub fn Integer(value: i32) -> Self

pub fn Double(value: f64) -> Self

pub fn Symbol(value: Interned) -> Self

pub fn BigInteger(value: Gc<BigInt>) -> Self

pub fn String(value: Gc<String>) -> Self

pub fn is_value_ptr<T: HasPointerTag>(&self) -> bool

pub fn as_value_ptr<T: HasPointerTag>(&self) -> Option<Gc<T>>

pub fn as_array(self) -> Option<VecValue>

Returns this value as an array, if such is its type.

pub fn as_block(self) -> Option<Gc<Block>>

Returns this value as a block, if such is its type.

pub fn as_class(self) -> Option<Gc<Class>>

Returns this value as a class, if such is its type.

pub fn as_instance(self) -> Option<Gc<Instance>>

Returns this value as an instance, if such is its type.

pub fn as_invokable(self) -> Option<Gc<Method>>

Returns this value as an invokable, if such is its type.

pub fn class(&self, universe: &Universe) -> Gc<Class>

Get the class of the current value.

pub fn lookup_method( &self, universe: &Universe, signature: Interned, ) -> Option<Gc<Method>>

Search for a given method for this value.

pub fn to_string(&self, universe: &Universe) -> String

Get the string representation of this value.

impl Value

pub fn Array(value: VecValue) -> Self

pub fn Block(value: Gc<Block>) -> Self

pub fn Class(value: Gc<Class>) -> Self

pub fn Instance(value: Gc<Instance>) -> Self

pub fn Invokable(value: Gc<Method>) -> Self