som_interpreter_bc::value::value_enum

Enum ValueEnum

pub enum ValueEnum {
    Nil,
    Boolean(bool),
    Integer(i32),
    BigInteger(Gc<BigInt>),
    Double(f64),
    Symbol(Interned),
    String(Gc<String>),
    Array(Gc<Vec<ValueEnum>>),
    Block(Gc<Block>),
    Instance(Gc<Instance>),
    Class(Gc<Class>),
    Invokable(Gc<Method>),
}

Represents an SOM value as an enum.

Variants

Nil

The nil value.

Boolean(bool)

A boolean value (true or false).

Integer(i32)

An integer value.

BigInteger(Gc<BigInt>)

A big integer value (arbitrarily big).

Double(f64)

An floating-point value.

Symbol(Interned)

An interned symbol value.

String(Gc<String>)

A string value.

Array(Gc<Vec<ValueEnum>>)

An array of values.

Block(Gc<Block>)

A block value, ready to be evaluated.

Instance(Gc<Instance>)

A generic (non-primitive) class instance.

Class(Gc<Class>)

A bare class object.

Invokable(Gc<Method>)

A bare invokable.

Implementations

impl ValueEnum

pub fn class(&self, universe: &Universe) -> Gc<Class>

Get the class of the current value.

pub fn lookup_method( &self, universe: &Universe, signature: Interned, ) -> Option<Gc<Method>>

Search for a given method for this value.

pub fn lookup_local(&self, idx: usize) -> Self

Search for a local binding within this value.

pub fn assign_local(&mut self, idx: usize, value: Self)

Assign a value to a local binding within this value.

pub fn to_string(&self, universe: &Universe) -> String

Get the string representation of this value.

impl ValueEnum

pub const NIL: Self = ValueEnum::Nil

The nil value.

pub const TRUE: Self

The boolean true value.

pub const FALSE: Self

The boolean false value.

pub const INTEGER_ZERO: Self

The integer 0 value.

pub const INTEGER_ONE: Self

The integer 1 value.

pub fn is_big_integer(&self) -> bool

Returns whether this value is a big integer.

pub fn is_string(&self) -> bool

Returns whether this value is a string.

pub fn is_array(&self) -> bool

Returns whether this value is an array.

pub fn is_block(&self) -> bool

Returns whether this value is a block.

pub fn is_class(&self) -> bool

Returns whether this value is a class.

pub fn is_instance(&self) -> bool

Returns whether this value is an instance.

pub fn is_invokable(&self) -> bool

Returns whether this value is an invokable.

pub fn is_nil(&self) -> bool

Returns whether this value is `nil``.

pub fn is_integer(&self) -> bool

Returns whether this value is an integer.

pub fn is_boolean(&self) -> bool

Returns whether this value is a boolean.

pub fn is_boolean_true(&self) -> bool

Returns whether or not it’s a boolean corresponding to true. NB: does NOT check if the type actually is a boolean.

pub fn is_boolean_false(&self) -> bool

Returns whether or not it’s a boolean corresponding to false. NB: does NOT check if the type actually is a boolean.

pub fn is_symbol(&self) -> bool

Returns whether this value is a symbol.

pub fn is_double(&self) -> bool

Returns whether this value is a double.

pub fn as_big_integer(&self) -> Option<Gc<BigInt>>

Returns this value as a big integer, if such is its type.

pub fn as_string(&self) -> Option<Gc<String>>

Returns this value as a string, if such is its type.

pub fn as_array(&self) -> Option<Gc<Vec<ValueEnum>>>

Returns this value as an array, if such is its type.

pub fn as_block(&self) -> Option<Gc<Block>>

Returns this value as a block, if such is its type.

pub fn as_class(&self) -> Option<Gc<Class>>

Returns this value as a class, if such is its type.

pub fn as_instance(&self) -> Option<Gc<Instance>>

Returns this value as an instance, if such is its type.

pub fn as_invokable(&self) -> Option<Gc<Method>>

Returns this value as an invokable, if such is its type.

pub fn as_integer(&self) -> Option<i32>

Returns this value as an integer, if such is its type.

pub fn as_boolean(&self) -> Option<bool>

Returns this value as a boolean, if such is its type.

pub fn as_symbol(&self) -> Option<Interned>

Returns this value as a symbol, if such is its type.

pub fn as_double(&self) -> Option<f64>

Returns this value as a double, if such is its type.

pub fn new_boolean(value: bool) -> Self

Returns a new boolean value.

pub fn new_integer(value: i32) -> Self

Returns a new integer value.

pub fn new_double(value: f64) -> Self

Returns a new double value.

pub fn new_symbol(value: Interned) -> Self

Returns a new symbol value.

pub fn new_big_integer(value: Gc<BigInt>) -> Self

Returns a new big integer value.

pub fn new_string(value: Gc<String>) -> Self

Returns a new string value.

pub fn new_array(value: Gc<Vec<ValueEnum>>) -> Self

Returns a new array value.

pub fn new_block(value: Gc<Block>) -> Self

Returns a new block value.

pub fn new_class(value: Gc<Class>) -> Self

Returns a new class value.

pub fn new_instance(value: Gc<Instance>) -> Self

Returns a new instance value.

pub fn new_invokable(value: Gc<Method>) -> Self

Returns a new invokable value.