som_interpreter_bc::vm_objects::frame

Struct Frame

pub struct Frame {
    pub prev_frame: Gc<Frame>,
    pub context: Gc<MethodInfo>,
    pub bytecode_idx: u16,
    pub args_marker: PhantomData<[Value]>,
    pub locals_marker: PhantomData<[Value]>,
}

Represents a stack frame.

Fields
prev_frame: Gc<Frame>

The previous frame. Frames are handled as a linked list

context: Gc<MethodInfo>

The method the frame is associated with.

bytecode_idx: u16

Bytecode index, needed to know where to resume execution when returning to parent frames.

args_marker: PhantomData<[Value]>

markers. we don’t use them directly. it’s mostly a reminder that the struct looks different in memory… not the cleanest but not sure how else to go about it

locals_marker: PhantomData<[Value]>
Implementations

impl Frame

pub fn alloc_initial_method( init_method: Gc<MethodInfo>, args: &[Value], gc_interface: &mut GCInterface, ) -> Gc<Frame>

Allocates the very first frame, for the initialize: call and tests. TODO: it’s a bit of an awkward method, so maybe check if it is strictly necessary. Historically was needed because normal method frames would need the stack from the previous frame, but now the stack is global.

pub fn get_true_size(nbr_args: u8, nbr_locals: u8) -> usize

Returns the true size of the Frame, counting the extra memory needed for its locals/arguments.

pub fn get_bytecodes(&self) -> &Vec<u8>

pub unsafe fn get_inline_cache_entry( &mut self, idx: usize, ) -> &mut Option<CacheEntry>

Safety

So long as idx is a bytecode_idx, it’s valid, since there’s as many entries as there are bytecode. Otherwise, it could break.

pub fn get_nbr_args(&self) -> u8

pub fn get_nbr_locals(&self) -> u8

pub fn lookup_local(&self, idx: usize) -> &Value

Search for a local binding. This function, and its friends, is kinda ugly. That’s what you get for using self-referential pointers.

pub fn assign_local(&mut self, idx: usize, value: Value)

Assign to a local binding.

pub fn lookup_argument(&self, idx: usize) -> &Value

pub fn assign_arg(&mut self, idx: usize, value: Value)

Assign to an argument.

pub fn lookup_constant(&self, idx: usize) -> &Literal

pub fn nth_frame_back(current_frame: &Gc<Frame>, n: u8) -> Gc<Frame>

Returns the nth frame back in the frame list, given n and the current frame. Walks the frame list not using the back pointer prev_frame, but by looking up the previous frame associated with each block.