Crate som_interpreter_bc

This is the interpreter for the Simple Object Machine.

Modules

block
Facilities for manipulating blocks.

class
Facilities for manipulating classes.

compiler
Facilities for compiling code into bytecode.

frame
Facilities for manipulating stack frames.

hashcode
Facilities for manipulating values.

instance
Facilities for manipulating class instances.

interner
Facilities for string interning.

interpreter
The interpreter's main data structure.

method
Facilities for manipulating class methods.

primitives
Definitions for all supported primitives.

universe
The collection of all known SOM objects during execution.

value
Facilities for manipulating values.

Macros

expect_args
Macro for checking and destructuring arguments passed to primitives.

reverse

Type Definitions

SOMRef
A strong and owning reference to an object.

SOMWeakRef
A weak reference to an object.