som_interpreter_bc::vm_objects::frame

Struct Frame

pub struct Frame {
    pub prev_frame: Gc<Frame>,
    pub current_context: Gc<Method>,
    pub bytecode_idx: u16,
    pub stack_ptr: u8,
    pub nbr_args: u8,
    pub max_stack_size: u8,
    pub stack_marker: PhantomData<[Value]>,
    pub args_marker: PhantomData<[Value]>,
    pub locals_marker: PhantomData<[Value]>,
}

Represents a stack frame.

Fields
prev_frame: Gc<Frame>

The previous frame. Frames are handled as a linked list

current_context: Gc<Method>

The method the execution context currently is in. Interestingly, this is a Gc and not a pointer to MethodInfo, what we really need (it’s never a primitive). In fact, we induce (minimal) runtime overhead by having to fetch the info from the Method enum regularly. So why do we do things that way? Because of moving GC. If we have a pointer to a MethodInfo, that’s an inner pointer to a Method object. So when GC moves the frame, it can’t update that pointer. It could update if Gc was a thing. And it was, at some point, and it turned out that really broke things (not sure why, I assume because MMTk didn’t play well with a Gc that had a Gc variant)

bytecode_idx: u16

Bytecode index.

stack_ptr: u8

Stack pointer/index. Points to the NEXT element that can be written to the stack; Alternatively, can be seen as number of elements on the stack

nbr_args: u8

It’s also stored in the current context, but we keep it here for faster access since we need it to calculate the offset to local variables

max_stack_size: u8

Needed for similar reasons as the number of arguments, for easier access to args and locals.

stack_marker: PhantomData<[Value]>

markers. we don’t use them directly. it’s mostly a reminder that the struct looks different in memory… not the cleanest but not sure how else to go about it

args_marker: PhantomData<[Value]>locals_marker: PhantomData<[Value]>
Implementations

impl Frame

pub fn alloc_from_block( nbr_args: usize, prev_frame: &mut Gc<Frame>, gc_interface: &mut GCInterface, ) -> Gc<Frame>

Allocates a frame for a block. We assume that the block is on the stack of the previous frame, as is the case when calling the primitive functions that create new blocks. We do this to make sure it’s reachable during GC.

pub fn alloc_initial_method( init_method: Gc<Method>, args: &[Value], gc_interface: &mut GCInterface, ) -> Gc<Frame>

Allocates the very first frame, for the initialize: call and tests. Special-cased because the normal case pushes the previous value on the previous frame’s stack for it to be reachable: we have no previous frame in some cases, so we can’t.

pub fn get_true_size( max_stack_size: usize, nbr_args: usize, nbr_locals: usize, ) -> usize

Returns the true size of the Frame, counting the extra memory needed for its stack/locals/arguments.

pub fn get_bytecode_ptr(&self) -> *const Vec<Bytecode>

pub fn get_max_stack_size(&self) -> usize

pub unsafe fn get_inline_cache_entry( &mut self, idx: usize, ) -> &mut Option<CacheEntry>

Safety

So long as idx is a bytecode_idx, it’s valid, since there’s as many entries as there are bytecode. Otherwise, it could break.

pub fn get_nbr_args(&self) -> usize

pub fn get_nbr_locals(&self) -> usize

pub fn lookup_local(&self, idx: usize) -> &Value

Search for a local binding. This function, and its friends, is kinda ugly. That’s what you get for using self-referential pointers. NOTE: fetching locals like this instead of storing a pointer in the frame to locals and arguments respectively, is a very minor slowdown. But moving GC and self-referential pointers led to some ugly bugs.. I think storing self-referential pointers into UnsafeCells might be enough to fix things though - TODO.

pub fn assign_local(&mut self, idx: usize, value: Value)

Assign to a local binding.

pub fn lookup_argument(&self, idx: usize) -> &Value

pub fn assign_arg(&mut self, idx: usize, value: Value)

Assign to an argument.

pub fn lookup_constant(&self, idx: usize) -> &Literal

pub fn nth_frame_back(current_frame: &Gc<Frame>, n: u8) -> Gc<Frame>

pub fn nth_frame_back_through_frame_list( current_frame: &Gc<Frame>, n: u8, ) -> Gc<Frame>

nth_frame_back but through prev_frame ptr. TODO: clarify why different implems are needed

pub unsafe fn nth_stack(&self, n: u8) -> &Value

Gets the nth element from the stack (not in reverse order - “3” yields the 3rd element from the bottom, not the top)

Safety

The caller needs to ensure this is a valid stack value. That means not outside the stack’s maximum size, and not pointing to an uninitialized value.

pub unsafe fn nth_stack_mut(&mut self, n: u8) -> &mut Value

Gets the nth element from the stack mutably (not in reverse order - “3” yields the 3rd element from the bottom, not the top)

Safety

The caller needs to ensure this is a valid stack value. That means not outside the stack’s maximum size, and not pointing to an uninitialized value.

pub fn stack_push(&mut self, value: Value)

pub fn stack_pop(&mut self) -> Value

pub fn stack_last(&self) -> &Value

pub fn stack_last_mut(&mut self) -> &mut Value

pub fn stack_nth_back(&self, n: usize) -> &Value

pub fn stack_n_last_elements(&self, n: usize) -> &'static [Value]

pub fn remove_n_last_elements(&mut self, n: usize)

pub fn stack_len(&self) -> usize

Gets the total number of elements on the stack. Only used for debugging.